delete temp plugins
