+++ src/thunder/include/plugin_registry.h
@@ -15,8 +15,6 @@
 #include "plugins/loaders/dyn_loader.h"
 #include "plugins/loaders/soft_loader.h"
 #include "plugins/loaders/urdf_loader.h"
-#include "plugins/loaders/kinTree_loader.h"
-#include "plugins/loaders/dynTree_loader.h"
 
 // - Builders - //
 #include "plugins/builders/kin_builder.h"
@@ -24,8 +22,6 @@
 #include "plugins/builders/reg_builder.h"
 #include "plugins/builders/soft_builder.h"
 #include "plugins/builders/example_builder.h"
-#include "plugins/builders/kinTree_builder.h"
-#include "plugins/builders/dynTree_builder.h"
 
 // - Generators - //
 #include "plugins/generators/robot_generator.h"
@@ -43,8 +39,6 @@ inline const std::map<std::string, std::shared_ptr<BaseLoader>> LOADERS = {
 	{"dyn_loader", std::make_shared<DynLoader>()},
 	{"soft_loader", std::make_shared<SoftLoader>()},
 	{"urdf_loader", std::make_shared<UrdfLoader>()},
-	{"kinTree_loader", std::make_shared<KinTreeLoader>()},
-	{"dynTree_loader", std::make_shared<DynTreeLoader>()},
 };
 
 /**
@@ -57,8 +51,6 @@ inline const std::map<std::string, std::shared_ptr<BaseBuilder>> POPULATORS = {
 	{"reg_builder", std::make_shared<RegBuilder>()},
 	{"soft_builder", std::make_shared<SoftBuilder>()},
 	{"example_builder", std::make_shared<ExampleBuilder>()},
-	{"kinTree_builder", std::make_shared<KinTreeBuilder>()},
-	{"dynTree_builder", std::make_shared<DynTreeBuilder>()},
 };
 
 /**
