+++ src/thunder/include/plugins/loaders/kin_loader.h
@@ -1,18 +1,18 @@
-#ifndef KIN_LOADER_H
-#define KIN_LOADER_H
+#ifndef KINTREE_LOADER_H
+#define KINTREE_LOADER_H
 
 #include "plugin_interfaces.h"
 
 
 namespace thunder_ns {
 
-	class KinLoader : public BaseLoader {
+	class KinTreeLoader : public BaseLoader {
 		
 		private:
 			std::string robot_name;
 
 		public:
-			KinLoader() : BaseLoader("Kinematics Loader", "Load the robot kinematics and structure.") {}
+			KinTreeLoader() : BaseLoader("Kinematics Loader", "Load the robot kinematics and structure, with tree support.") {}
 
 			std::shared_ptr<Robot> load(std::shared_ptr<Robot> robot_ptr) override;
 
@@ -21,4 +21,4 @@ namespace thunder_ns {
 
 } // namespace thunder_ns
 
-#endif // KIN_LOADER_H
+#endif // KINTREE_LOADER_H
